Remove set-course-image handling
This is a remnant from before this element and the image-selector were broken apart, and the "when a course image is selected, close the overlay" behaviour is specific to that use case. Instead, it should be up to consumers to close the overlay at the correct time, either by calling .close(), or by firing a d2l-simple-overlay-close-button-clicked event.