Fix scrollable area height
I feel like this has probably always been here, but I guess I'd just never noticed it - if you have large (scrollable) content, the bottom of the scroll bar is currently cut off, along with the bottom few pixels of content. It seems that the header is actually 150px in height, not 100.